Update all files of the screen , add MusicManager.kt

+++ composeApp/src/commonMain/kotlin/org/example/project/Score/Score.kt
@@ -6,8 +6,7 @@ import kotlinx.serialization.Transient
 
 @Serializable
 data class Score (
-    val id: Long? = null,
-
+    val id: Int? = null,
     @SerialName("name")
     val name: String,
